@@ -11,27 +11,28 @@
     "url": "https://github.com/idea2app/React-MobX-Ant-Design-ts/issues"
   },
   "dependencies": {
-    "@ant-design/icons": "^4.6.2",
-    "antd": "^4.16.0",
+    "@ant-design/icons": "^4.7.0",
+    "antd": "^4.18.2",
+    "browser-unhandled-rejection": "^1.0.2",
     "mobx": "^5.15.7",
     "mobx-react": "^6.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0"
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "web-utility": "^2.9.6"
   },
   "devDependencies": {
-    "@octokit/openapi-types": "^7.2.3",
-    "@snowpack/plugin-react-refresh": "^2.5.0",
-    "@snowpack/plugin-typescript": "^1.2.1",
-    "@types/react-dom": "^17.0.5",
-    "@types/react-router": "^5.1.14",
-    "@types/react-router-dom": "^5.1.7",
-    "husky": "^6.0.0",
-    "lint-staged": "^11.0.0",
-    "prettier": "^2.3.0",
-    "snowpack": "^3.3.7",
-    "typescript": "^4.3.2"
+    "@octokit/openapi-types": "^11.2.0",
+    "@parcel/packager-raw-url": "^2.0.1",
+    "@parcel/transformer-webmanifest": "^2.0.1",
+    "@types/react-dom": "^17.0.11",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.1.4",
+    "parcel": "^2.0.1",
+    "prettier": "^2.5.1",
+    "typescript": "~4.4.4",
+    "workbox-cli": "^6.4.2"
   },
   "prettier": {
     "singleQuote": true,
@@ -44,7 +45,8 @@
   "scripts": {
     "prepare": "husky install",
     "test": "lint-staged",
-    "start": "snowpack dev",
-    "build": "rm -rf build/  &&  snowpack build"
+    "start": "parcel src/index.html --open",
+    "pack": "parcel build src/index.html --public-url .",
+    "build": "rm -rf dist/  &&  npm run pack  &&  workbox generateSW"
   }
 }
