@@ -1,12 +1,11 @@
-import React from "react";
-import { HashRouter, Route, Switch } from "react-router-dom";
+import { HashRouter, Routes, Route } from 'react-router-dom';
 
-import { HomePage } from "./Home";
+import { HomePage } from './Home';
 
 export default () => (
   <HashRouter>
-    <Switch>
-      <Route path="/" exact component={HomePage} />
-    </Switch>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+    </Routes>
   </HashRouter>
 );
